Remove zero values from ScalaType.
We used them as default values in MemoryDriver but this is not necessary
for compatibility with SQL databases. Inserting into a row without an
explicit value or an explicit default defined in the schema is always
equivalent to inserting NULL. On a real database it leads to a
constraint violation if the column in question is not nullable. With
this change, MemoryDriver also uses null as the default for all data
types. Constraints are not enforced (in any scenario in MemoryDriver).